+++ db/migrate/20170216115256_create_chapters.rb
@@ -1,8 +1,6 @@
 class CreateChapters < ActiveRecord::Migration[5.0]
   def change
-    if table_exists?(:chapters)
-      drop_table(:chapters)
-    end
+    drop_table(:chapters) if table_exists?(:chapters)
 
     create_table :chapters do |t|
       t.references :book, foreign_key: true
